Safari-611.1.10/LayoutTests/fast/box-shadow/hit-test-box-shadow-and-margin-on-zero-height-clipping-container-expected.txt

PASS: target element is #container-with-height
